requirements.txt
asgiref==3.2.10
Django==3.1.2
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
gunicorn
pillow
django-filter
whitenoise
